VK-GL-CTS issue: 5498

Components: Framework, Vulkan

Affects:
dEQP-VK.draw.*.instanced.*
dEQP-VK.compute.*.indirect_dispatch.*
dEQP-VK.conditional_rendering.draw_clear.draw.*
dEQP-VK.query_pool.occlusion_query.*
dEQP-VK.query_pool.statistics_query.*
dEQP-VK.api.copy_and_blit.*
dEQP-VK.api.device_address.copy_flags.*
dEQP-VK.ray_query.*
dEQP-VK.ray_tracing_pipeline.*
dEQP-VK.api.fill_and_update_buffer.*
dEQP-VK.mesh_shader.ext.api.*
dEQP-VK.robustness.index_access.*
dEQP-VK.robustness.bind_index_buffer2.*
dEQP-VK.transform_feedback.*

+++ external/openglcts/modules/gl/gl4cDirectStateAccessBuffersTests.cpp
@@ -794,7 +794,7 @@ void ClearTest::LogFail(bool use_sub_data, glw::GLenum internalformat, const T *
     {
         std::stringstream number;
 
-        number << data[i];
+        number << (GLint)data[i];
 
         the_log.append(number.str());
 
@@ -810,7 +810,7 @@ void ClearTest::LogFail(bool use_sub_data, glw::GLenum internalformat, const T *
     {
         std::stringstream number;
 
-        number << reference[i];
+        number << (GLint)reference[i];
 
         the_log.append(number.str());
